--- src/app/discover/page.tsx
@@ -1,12 +1,0 @@
-import Hero from "@/components/sections/hero";
-import Features from "@/components/sections/features";
-
-
-export default function Discover() {
-  return (
-    <main>
-      <Hero />
-      <Features />
-    </main>
-  );
-}
