+++ src/app/page.tsx
@@ -2,12 +2,14 @@
 
 import HeroSection from "@/components/HeroSection";
 import FaqSection from "@/components/FaqSection";
+import Discover from "@/components/DiscoverSection";
 
 export default function Home() {
   return (
     <>
       <HeroSection />
       <FaqSection />
+      <Discover />
     </>
   );
 }
