--- src/components/sections/hero.tsx
@@ -1,13 +1,0 @@
-export default function Hero() {
-    return (
-      <section className="text-center dark:bg-black pt-16 px-6">
-        <h2 className="text-4xl font-bold">
-          Discover the Power of <span className="text-blue-600 dark:text-blue-400">SafeTrust</span>
-        </h2>
-        <p className="text-gray-600 dark:text-gray-400 mt-3 text-lg max-w-2xl mx-auto">
-          SafeTrust is engineered with cutting-edge blockchain technology designed to revolutionize
-          the way you handle P2P transactions, secure deposits, and manage digital assets.
-        </p>
-      </section>
-    );
-  }
